Move ProjectForm out of Submitter.
