+++ client/src/app/project-form/wrangler-layout.ts
@@ -1,4 +1,4 @@
-import {MetadataFormLayout} from '../../metadata-schema-form/models/metadata-form-layout';
+import {MetadataFormLayout} from '../metadata-schema-form/models/metadata-form-layout';
 
 export const wranglerLayout: MetadataFormLayout = {
   tabs: [
